+++ ephys/spiketrains.py
